Split db helper functions into separate files, add helpers object to definitions
