+++ src/db/helpers/index.js
@@ -1,149 +1,11 @@
 const bcrypt = require('bcrypt');
 
-const { getWeekBounds } = require('../../helpers')
+module.exports = function (db, helpers) {
 
-module.exports = function (db) {
-
-  const getUsernameById = id => {
-    return db.query(`SELECT username FROM users WHERE id = $1`, [id])
-             .then(rows => {
-               if (rows[0]) {
-                 return rows[0].username
-               } else {
-                 return null;
-               }
-             })
-  }
-
-  const getIdByUsername = username => {
-    return db.query(`SELECT id FROM users WHERE username = $1`, [username])
-             .then(rows => {
-               if (rows[0]) {
-                 return rows[0].id;
-               } else {
-                 return null;
-               }
-             })
-  }
-
-  // The user parameter must have two keys: "username" and "hashed_password"
-  const addUser = (username, rawPassword) => {
-    const saltRounds = 10;
-    return bcrypt.hash(rawPassword, saltRounds)
-                 .then(hashed_password => db.insert(`users`, {username, hashed_password}))
-                 .then(rows => rows[0])
-  }
-
-  const validLogin = (username, rawPassword) => {
-    // Find the hashed password matching the username
-    return db.query(`SELECT hashed_password FROM users WHERE username = $1`, [username])
-             .then(rows => {
-               // If there's a matching user, check the password and return result
-               if (rows[0]) {
-                const { hashed_password } = rows[0];
-                return bcrypt.compare(rawPassword, hashed_password);
-               } else {
-                // If there's no result, return false
-                return false;
-               }
-             });
-  }
-
-  // Defaults to the current week, but can accept any target date.
-  const getWeeklyBlocksByUser = (userId, targetDate = new Date()) => {
-    // Get Sunday and Saturday for given week.
-    const { lastSunday, nextSaturday } = getWeekBounds(targetDate);
-    console.log(`looking up blocks for user id ${userId} between ${lastSunday.toISOString()} and ${nextSaturday.toISOString()}`);
-
-    return db.query(`
-      SELECT blocks.*, projects.title FROM blocks
-        JOIN projects
-        ON blocks.project_id = projects.id
-      WHERE blocks.user_id = $1
-      AND start_time BETWEEN $2 AND $3
-      ORDER BY start_time
-      `, [userId, lastSunday.toISOString(), nextSaturday.toISOString()])
-  }
-
-  const getProjectsByUser = userId => {
-    return db.query(`
-      SELECT * FROM projects
-      WHERE user_id = $1
-      `, [userId])
-  }
-
-  const addProject = (userId, title) => {
-    const project = { user_id: userId, title };
-    return db.insert('projects', project)
-             .then(rows => rows[0]);
-  }
-
-  const updateProjectTitle = (userId, projectId, title) => {
-    return db.query(`
-      UPDATE projects SET title = $1
-      WHERE id = $2
-      AND user_id = $3
-      RETURNING *
-    `, [title, projectId, userId]);
-  }
-
-  // Creates a single new session in the DB and then returns the data about that session.
-  const startSession = sessionData => {
-    return db.insert('sessions', sessionData)
-             .then(rows => rows[0])
-  }
-
-  // "Stops" a session and returns the stopped session.
-  const stopSession = (userId, sessionId) => {
-    const nowStr = (new Date()).toISOString();
-    console.log(`Stopping session at: ${nowStr}`)
-    return db.query(`
-      UPDATE sessions SET end_time = $1
-      WHERE id = $2
-      AND user_id = $3
-      RETURNING *
-    `, [nowStr, sessionId, userId]);
-  }
-
-  const deleteSession = (userId, sessionId) => {
-    return db.query(`
-    DELETE FROM sessions
-    WHERE user_id = $1 AND id = $2
-    RETURNING *
-    `, [userId, sessionId])
-  }
-
-  // Updates a session with new start/end times.
-  const updateSession = sessionData => {
-    const { session_id, start_time, end_time } = sessionData
-    return db.update('sessions', session_id, {start_time, end_time});
-  }
-
-  const getWeeklySessions = (userId, targetDate = new Date()) => {
-    const { lastSunday, nextSaturday } = getWeekBounds(targetDate);
-    console.log(lastSunday.toISOString())
-    return db.query(`
-      SELECT sessions.*, projects.title
-      FROM sessions
-      JOIN projects
-        ON projects.id = sessions.project_id
-      WHERE sessions.user_id=$1
-      AND sessions.start_time >= $2
-      AND sessions.end_time <= $3
-      ORDER BY sessions.project_id DESC, sessions.start_time DESC
-      `, [userId, lastSunday.toISOString(), nextSaturday.toISOString()]);
-  }
-
-  // Get the most recent unfinished session for this user.
-  const getCurrentSession = (userId) => {
-    return db.query(`
-      SELECT * FROM sessions
-      WHERE user_id=$1
-      AND end_time IS NULL
-      ORDER BY start_time DESC
-      LIMIT 1
-      `, [userId]);
-  }
+  const userHelpers = require('./userHelpers')(db, helpers)
+  const blockHelpers = require('./blockHelpers')(db, helpers)
+  const projectHelpers = require('./projectHelpers')(db, helpers)
+  const sessionHelpers = require('./sessionHelpers')(db, helpers)
 
   // getWeeklyReport finds every project that:
   // 1. Belongs to this user
@@ -154,7 +16,7 @@ module.exports = function (db) {
   // NOTE: "EXTRACT EPOCH" is used to turn the time interval into a number of seconds.
   // That way we can calculate hours on the frontend.
   const getWeeklyReport = (userId, targetDate = new Date()) => {
-    const { lastSunday, nextSaturday } = getWeekBounds(targetDate);
+    const { lastSunday, nextSaturday } = helpers.getWeekBounds(targetDate);
 
     return db.query(`
       SELECT
@@ -182,27 +44,13 @@ module.exports = function (db) {
       `, [userId, lastSunday.toISOString(), nextSaturday.toISOString()]);
   }
 
-  const addBlock = blockData => {
-    return db.insert('blocks', blockData)
-             .then(rows => rows[0])
-  }
+
 
   return {
-    getUsernameById,
-    addUser,
-    validLogin,
-    getIdByUsername,
-    getWeeklyBlocksByUser,
-    getProjectsByUser,
-    addProject,
-    updateProjectTitle,
-    startSession,
-    stopSession,
-    updateSession,
-    deleteSession,
-    getWeeklySessions,
-    getCurrentSession,
-    getWeeklyReport,
-    addBlock
+    ...userHelpers,
+    ...blockHelpers,
+    ...projectHelpers,
+    ...sessionHelpers,
+    getWeeklyReport
   }
 }
